split label parsing tests to own file
